+++ content/register_content_modules.cpp
@@ -35,6 +35,7 @@ namespace chronon3d::content::certification { void register_cert_title_compositi
                                                void register_cert_lower_third_compositions(CompositionRegistry&);
                                                void register_cert_long_text_compositions(CompositionRegistry&);
                                                void register_cert_multilingual_compositions(CompositionRegistry&); }
+namespace chronon3d::content::launches    { void register_launches_compositions(CompositionRegistry&); }
 
 namespace chronon3d {
 
@@ -68,6 +69,7 @@ public:
         content::certification::register_cert_multilingual_compositions(ctx.compositions);
         content::text_placement::register_text_placement_compositions(ctx.compositions);
         content::sequence_v2::register_sequence_v2_compositions(ctx.compositions);
+        content::launches::register_launches_compositions(ctx.compositions);
     }
 };